+++ src/EulerHS/Core/Types/MySQL.hs
@@ -1,5 +1,6 @@
-{-# LANGUAGE DerivingStrategies #-}
+{-# OPTIONS_GHC -Werror #-}
 {-# LANGUAGE DeriveAnyClass     #-}
+{-# LANGUAGE DerivingStrategies #-}
 
 module EulerHS.Core.Types.MySQL
   (
@@ -7,6 +8,7 @@ module EulerHS.Core.Types.MySQL
     -- ** Types
     MySQLConfig(..)
   , MySqlOption(..)
+  , MySQLCharset(..)
     -- ** Methods
   , createMySQLConn
   , closeMySQLConn
@@ -14,12 +16,13 @@ module EulerHS.Core.Types.MySQL
   , defaultMySQLConfig
   ) where
 
-import Prelude
-import Data.Word (Word16)
-import Data.Aeson (ToJSON, FromJSON)
-import GHC.Generics (Generic)
-import Database.MySQL.Base (MySQLConn, close, ConnectInfo (..), connect, defaultConnectInfoMB4)
-import Data.ByteString.UTF8 (fromString)
+import           Data.Aeson (FromJSON, ToJSON)
+import           Data.ByteString.UTF8 (fromString)
+import           Data.Word (Word16, Word8)
+import           Database.MySQL.Base (ConnectInfo (..), MySQLConn, close,
+                                      connect)
+import           GHC.Generics (Generic)
+import           Prelude
 
 data MySqlProtocol
   = TCP
@@ -72,6 +75,35 @@ data SSLInfo = SSLInfo
   deriving stock (Show, Eq, Ord, Generic)
   deriving anyclass (ToJSON, FromJSON)
 
+-- | Describes the character set to be used with the database. This also
+-- includes collation information.
+--
+-- Currently, only a limited number of these are provided.
+--
+-- /See also:/ [MySQL documentation on character
+-- sets](https://dev.mysql.com/doc/refman/5.7/en/charset-mysql.html)
+--
+-- @since 2.0.3.0
+data MySQLCharset =
+  -- | Corresponds to the @latin1@ character set, with the @latin1_swedish_ci@
+  -- collation.
+  --
+  -- @since 2.0.3.0
+  Latin1
+  -- | Corresponds to the @utf8@ character set, with the @utf8_general_ci@
+  -- collation.
+  --
+  -- @since 2.0.3.0
+  | UTF8General
+  -- | Corresponds to the @utf8mb@ character set, with the @unicode_ci@
+  -- collation.
+  --
+  -- @since 2.0.3.0
+  | UTF8Full
+  deriving stock (Eq, Show, Ord, Generic)
+  deriving anyclass (ToJSON, FromJSON)
+
+-- | @since 2.0.3.0
 data MySQLConfig = MySQLConfig
   { connectHost     :: String
   , connectPort     :: Word16
@@ -81,10 +113,12 @@ data MySQLConfig = MySQLConfig
   , connectOptions  :: [MySqlOption]
   , connectPath     :: FilePath
   , connectSSL      :: Maybe SSLInfo
+  , connectCharset  :: !MySQLCharset -- ^ @since 2.0.3.0
   }
   deriving stock (Show, Eq, Ord, Generic)
   deriving anyclass (ToJSON, FromJSON)
 
+-- | @since 2.0.3.0
 defaultMySQLConfig :: MySQLConfig
 defaultMySQLConfig = MySQLConfig {
   connectHost = "localhost",
@@ -94,10 +128,13 @@ defaultMySQLConfig = MySQLConfig {
   connectDatabase = "test",
   connectOptions = [CharsetName "utf8"],
   connectPath = "",
-  connectSSL = Nothing
+  connectSSL = Nothing,
+  connectCharset = Latin1
   }
 
 -- | Connect with the given config to the database.
+--
+-- @since 2.0.3.0
 createMySQLConn :: MySQLConfig -> IO MySQLConn
 createMySQLConn conf = do
   let dbConf = ConnectInfo {
@@ -106,10 +143,18 @@ createMySQLConn conf = do
     ciDatabase = fromString . connectDatabase $ conf,
     ciUser = fromString . connectUser $ conf,
     ciPassword = fromString . connectPassword $ conf,
-    ciCharset = ciCharset defaultConnectInfoMB4
+    ciCharset = charsetToDBCharset . connectCharset $ conf
     }
   connect dbConf
 
 -- | Close the given connection.
 closeMySQLConn :: MySQLConn -> IO ()
 closeMySQLConn = close
+
+-- Helpers
+
+charsetToDBCharset :: MySQLCharset -> Word8
+charsetToDBCharset = \case
+  Latin1      -> 8
+  UTF8General -> 33
+  UTF8Full    -> 224